Discount % in float

+++ database/factories/UserFactory.php
@@ -157,7 +157,7 @@ $factory->define(IndianIra\ForgotPassword::class, function (Faker $faker) {
 $factory->define(IndianIra\Coupon::class, function (Faker $faker) {
     return [
         'code'             => strtoupper($faker->word) . mt_rand(1000, 9999),
-        'discount_percent' => round($faker->randomFloat(2, 0, 100), 2),
+        'discount_percent' => (float) round($faker->randomFloat(2, 0, 100), 2),
     ];
 });
